+++ test/test_callable_types.rb
@@ -158,10 +158,8 @@ class TestCallableTypes < Minitest::Test
         when "q" then [m, RatatuiRuby::Tea::Command.exit]
         else [m, nil]
         end
-      when Array
-        events << msg[0]
-        [m, nil]
       else
+        events << msg
         [m, nil]
       end
     end
@@ -193,10 +191,8 @@ class TestCallableTypes < Minitest::Test
         when "q" then [m, RatatuiRuby::Tea::Command.exit]
         else [m, nil]
         end
-      when Array
-        events << msg[0]
-        [m, nil]
       else
+        events << msg
         [m, nil]
       end
     end
@@ -232,10 +228,8 @@ class TestCallableTypes < Minitest::Test
         when "q" then [m, RatatuiRuby::Tea::Command.exit]
         else [m, nil]
         end
-      when Array
-        events << msg[0]
-        [m, nil]
       else
+        events << msg
         [m, nil]
       end
     end
@@ -273,10 +267,8 @@ class TestCallableTypes < Minitest::Test
         when "q" then [m, RatatuiRuby::Tea::Command.exit]
         else [m, nil]
         end
-      when Array
-        events << msg[0]
-        [m, nil]
       else
+        events << msg
         [m, nil]
       end
     end